changed min_price field type

+++ database/migrations/2020_01_17_192616_create_restaurateurs_table.php
@@ -25,7 +25,7 @@ class CreateRestaurateursTable extends Migration
             $table->string('vat_number', 11);
             $table->integer('max_delivery_time_slot')->unsigned();
             $table->float('delivery_cost', 4, 2)->default(1);
-            $table->integer('min_price')->default(1)->unsigned();
+            $table->float('min_price')->default(1)->unsigned();
             $table->string('image_path')->nullable();
             $table->unsignedBigInteger('shop_type_id');
             $table->string('device_id')->nullable();
